minor movments
