+++ package.json
@@ -20,6 +20,7 @@
     "react-slick": "^0.28.1",
     "react-three-fiber": "^6.0.13",
     "sass": "^1.49.9",
+    "styled-components": "^5.3.6",
     "three": "^0.142.0",
     "web-vitals": "^2.1.4"
   },
