--- public/3d/head/head.mtl
@@ -1,11 +1,0 @@
-newmtl 1
-Ns 0.1
-Ka 1.000000 1.000000 1.000000
-Kd 0.70000 0.70000 0.7000
-Ks 0.000000 0.000000 0.000000
-Ke 0.000000 0.000000 0.000000
-Ni 1.000000
-d 1.00000
-illum 2
-map_Kd 1647563332785_1.png
-
