+++ src/components/Background.js
@@ -1,0 +1,12 @@
+import '../css/background.css';
+import Me from './Me.js';
+
+function Empty() {
+    return (
+        <div id="background">
+            <Me />
+        </div>
+    );
+}
+
+export default Empty;
