+++ src/components/Button.js
@@ -1,0 +1,10 @@
+import '../css/button.css'
+
+export function Button(props) {
+
+    return (
+        <a className={props.className + " btn"} href={props.href}>{props.text}</a>
+    );
+}
+
+export default Button;
