--- src/js/Home.js
@@ -1,15 +1,0 @@
-import '../css/home.css';
-
-function Home() {
-    return (
-        <section id="home">
-            <center>
-                <div id="drop" class="drop">
-                    <a href="#aboutAnchor"><span></span>MORE</a>
-                </div>
-            </center>
-        </section>
-    );
-}
-
-export default Home;
